refactor: 调整项目结构-SettingsPage
