+++ src/pages/SettingsPage/components/BrowserSetting.tsx
@@ -18,18 +18,18 @@ import {
   AlertDialogHeader,
   AlertDialogTitle,
   AlertDialogTrigger,
-} from '../ui/alert-dialog'
-import { Button } from '../ui/button'
+} from '../../../components/ui/alert-dialog'
+import { Button } from '../../../components/ui/button'
 import {
   Card,
   CardContent,
   CardDescription,
   CardHeader,
   CardTitle,
-} from '../ui/card'
-import { Input } from '../ui/input'
-import { Label } from '../ui/label'
-import { Separator } from '../ui/separator'
+} from '../../../components/ui/card'
+import { Input } from '../../../components/ui/input'
+import { Label } from '../../../components/ui/label'
+import { Separator } from '../../../components/ui/separator'
 
 export function BrowserSetting() {
   const path = useCurrentChromeConfig(context => context.path)
